fix: bump authentication module `satellizer`
Using the v0.15.5-oodrive of satellizer which contains the to allow
authorizationEndpoint with query parameters as Oodrive works with
workspace concepts.

Link to the project that fixed the issue: https://github.com/l-lin/satellizer/

It's a forked project because the Satellizer maintainer no longer
maintains the project, thus no fix nor publication can be performed in
the original project repository.